requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
python-multipart